pyproject.toml
[project]
name = "webvulnscanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
]
